requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.0
click==8.1.7
coverage==7.4.4
Django==5.0.4
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.9.0
flake8==7.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
pytest==8.1.1
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
uritemplate==4.1.1
